+++ dakara_server/playlist/apps.py
@@ -1,19 +1,8 @@
 from internal.apps import DakaraConfig
+from playlist import signals  # noqa F401
 
 
 class PlaylistConfig(DakaraConfig):
     """Playlist app."""
 
     name = "playlist"
-
-    def ready_no_reload(self):
-        """Method called when app starts."""
-        from playlist.date_stop import check_date_stop_on_app_ready
-
-        check_date_stop_on_app_ready()
-
-    def ready_reload(self):
-        """Method called when app starts or restarts."""
-        from playlist.models import clean_channel_names
-
-        clean_channel_names()