+++ dakara_server/playlist/models.py
@@ -18,7 +18,7 @@ class PlaylistManager(OrderedModelManager):
 
     def get_playing(self):
         """Get the current playlist entry."""
-        playlist = self.filter(was_played=False, date_played__isnull=False)
+        playlist = self.filter(was_played=False, date_play__isnull=False)
 
         if not playlist:
             return None
@@ -33,15 +33,15 @@ class PlaylistManager(OrderedModelManager):
 
         return playlist.first()
 
-    def get_playlist(self):
+    def get_queuing(self):
         """Get the playlist of ongoing entries."""
         queryset = self.exclude(
-            models.Q(was_played=True) | models.Q(date_played__isnull=False)
+            models.Q(was_played=True) | models.Q(date_play__isnull=False)
         )
 
         return queryset
 
-    def get_playlist_played(self):
+    def get_played(self):
         """Get the playlist of passed entries."""
         playlist = self.filter(was_played=True)
 
@@ -62,10 +62,10 @@ class PlaylistManager(OrderedModelManager):
 
         else:
             # do not process a played entry
-            if self.get_playlist_played().filter(pk=entry_id):
+            if self.get_played().filter(pk=entry_id):
                 return None
 
-            playlist = self.get_playlist().exclude(pk=entry_id)
+            playlist = self.get_queuing().exclude(pk=entry_id)
 
         if not playlist:
             return None
@@ -83,7 +83,7 @@ class PlaylistEntry(OrderedModel):
     date_created = models.DateTimeField(auto_now_add=True)
     owner = models.ForeignKey(DakaraUser, null=False, on_delete=models.CASCADE)
     was_played = models.BooleanField(default=False, null=False)
-    date_played = models.DateTimeField(null=True)
+    date_play = models.DateTimeField(null=True)
 
     class Meta(OrderedModel.Meta):
         pass
@@ -98,7 +98,7 @@ class PlaylistEntry(OrderedModel):
             raise RuntimeError("A playlist entry is currently in play")
 
         # set the playlist entry
-        self.date_played = datetime.now(tz)
+        self.date_play = datetime.now(tz)
         self.save()
 
     def set_finished(self):